+++ src/main/kotlin/com/hwalaon/wezxro_server/global/security/principal/PrincipalDetailsService.kt
@@ -15,7 +15,7 @@ class PrincipalDetailsService(
 ): UserDetailsService {
     override fun loadUserByUsername(username: String?): UserDetails =
         if (!username.isNullOrBlank()) PrincipalDetails(
-            accountEntityRepository.findAllByUserIdAndStatusNot(username.toInt()).let {
+            accountEntityRepository.findAllByUserIdAndStatusNot(username.toLong()).let {
                 if (it == null) throw AccountNotFoundException()
                 accountMapper.toDomain(it)
             }
